fix(model): harden SenseNova rollout contracts
Normalize stochastic transitions into unit-noise coordinates, match official CFG reduction semantics, move prefix caches with conditions, and remove redundant per-branch and batch work while expanding contract tests.

+++ unirl/models/sensenova_u1/config.py
@@ -27,7 +27,7 @@ class SenseNovaU1PipelineConfig:
     device: Any = None
 
     autocast_precision: str = "bf16"
-    trajectory_precision: str = "fp32"
+    trajectory_precision: str = "bf16"
     logprob_precision: str = "fp32"
 
     timestep_shift: float = 3.0
